+++ config/routes.rb
@@ -1,14 +1,12 @@
 Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
+      mount_devise_token_auth_for 'User', at: 'auth', controllers: {
+        omniauth_callbacks: 'api/v1/auth/omniauth_callbacks'
+      }
       namespace :auth do
         resources :sessions, only: %i[index]
       end
-      mount_devise_token_auth_for 'User', at: 'auth', controllers: {
-        omniauth_callback: 'overrides/omniauth_callback'
-      }
-
-      resources :profiles, only: %i[index show edit update]
     end
   end
 end
